[Enhancement] limit max auto increment cache size (#21249) (#21251)
In current implementation, if the auto increment id counter is overflow, the insert/load must be fail.
And if we set the cache size too large, it will cause the overflow manually which is a problem here.
So the limitation must be set to prevent this situation. We can simply define the cache size as int instead of long

+++ fe/fe-core/src/main/java/com/starrocks/common/Config.java
@@ -2211,7 +2211,7 @@ public class Config extends ConfigBase {
      * default size of minimum cache size of auto increment id allocation
      **/
     @ConfField(mutable = true)
-    public static long auto_increment_cache_size = 100000;
+    public static int auto_increment_cache_size = 100000;
 
     /**
      * Enable the experimental temporary table feature
